Remove submenu and implement basic ui-library table stuff
