+++ admin/menu/class-menu.php
@@ -84,10 +84,6 @@ class WPSEO_Menu implements WPSEO_WordPress_Integration {
 				require_once WPSEO_PATH . 'admin/pages/social.php';
 				break;
 
-			case 'wpseo_indexables':
-				require_once WPSEO_PATH . 'admin/pages/indexables.php';
-				break;
-
 			case 'wpseo_licenses':
 				require_once WPSEO_PATH . 'admin/pages/licenses.php';
 				break;
